Bug fix. Items itemable_id is now unsigned big int

+++ database/migrations/2020_11_08_181924_create_invoice_items_table.php
@@ -20,7 +20,7 @@ class CreateInvoiceItemsTable extends Migration
             $table->integer('quantity');
             $table->decimal('price', 12);
 
-            $table->integer('itemable_id');
+            $table->unsignedBigInteger('itemable_id');
             $table->string('itemable_type');
 
             $table->timestamps();
